+++ tests/test_credentials.py
@@ -1,3 +1,14 @@
-from unittest.mock import MagicMock
+import pytest
+from httpx import AsyncClient
 
-from prefect import flow
+from prefect_databricks import DatabricksCredentials
+
+
+@pytest.mark.parametrize("token", [None, "token_value"])
+def test_databricks_credentials_get_client(token):
+    client = DatabricksCredentials(
+        databricks_instance="databricks_instance", token=token
+    ).get_client()
+    assert isinstance(client, AsyncClient)
+    if token is not None:
+        assert client.headers["authorization"] == "Bearer token_value"