+++ CHANGELOG.md
@@ -21,8 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.1.0
 
-Released on ????? ?th, 20??.
+Released on August 15th, 2022.
 
 ### Added
 
-- `task_name` task - [#1](https://github.com/PrefectHQ/prefect-databricks/pull/1)
+- `execute_endpoint` and `jobs*` tasks - [#1](https://github.com/PrefectHQ/prefect-databricks/pull/1)